Create .bashrc
Configurations were made to my .bashrc file for compiling and resetting my code files to my template file with ease.

The working of this is pretty self-explanatory. It should be noted that my C++ template is stored in the t.cpp file which is loaded in my main C++ files on invoking the prep command.